finished course

+++ resources/store.py
@@ -1,4 +1,5 @@
 from flask_restful import Resource
+from flask_jwt_extended import jwt_required
 from models.store import StoreModel
 
 
@@ -21,7 +22,11 @@ class Store(Resource):
 
         return store.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         store = StoreModel.find_by_name(name)
         if store:
             store.delete_from_db()
@@ -31,5 +36,5 @@ class Store(Resource):
 
 class StoreList(Resource):
     def get(self):
-        return {'stores': [store.json() for store in StoreModel.query.all()]}
+        return {'stores': [store.json() for store in StoreModel.find_all()]}
